// src/enums/notification.enum.ts
export enum NotificationType {
  INFO = 'info',
  SUCCESS = 'success',
  WARNING = 'warning',
  ERROR = 'error',
  POST_POST_NOW = 'post_post_now',
  SOCIAL_CONNECTED = 'social_connected',
  SYSTEM = 'system',
  USER_WELCOME = 'user_welcome',
  CHAT = 'chat',
}

export enum NotificationStatus {
  UNREAD = 'unread',
  READ = 'read',
  ARCHIVED = 'archived',
}